@@ -5,18 +5,15 @@ package org.georchestra.signalement.api.security;
 
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Enumeration;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.commons.lang3.StringUtils;
 import org.georchestra.signalement.core.dto.User;
 import org.georchestra.signalement.service.sm.UserService;
@@ -24,6 +21,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.util.AntPathMatcher;
+import org.springframework.web.filter.OncePerRequestFilter;
 
 /**
  * This filter is inspired from georchestra geowebcache project
@@ -31,7 +30,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
  * @author FNI18300
  *
  */
-public class PreAuthenticationFilter implements Filter {
+public class PreAuthenticationFilter extends OncePerRequestFilter {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(PreAuthenticationFilter.class);
 
@@ -45,41 +44,22 @@ public class PreAuthenticationFilter implements Filter {
 	public static final String SEC_LASTNAME = "sec-lastname";
 	public static final String SEC_FIRSTNAME = "sec-firstname";
 
-	private UserService userService;
+	// Controle des patterns des URL
+	private AntPathMatcher pathMatcher;
 
-	public PreAuthenticationFilter(UserService userService) {
+
+	// Liste des URL à exclure
+	private Collection<String> excludeUrlPatterns;
+	
+	private final UserService userService;
+
+	public PreAuthenticationFilter(final String[] excludeUrlPatterns, UserService userService) {
 		super();
 		this.userService = userService;
+		this.excludeUrlPatterns = Arrays.asList(excludeUrlPatterns);
+		this.pathMatcher = new AntPathMatcher();
 	}
 
-	@Override
-	public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-			throws IOException, ServletException {
-		if (request instanceof HttpServletRequest) {
-			HttpServletRequest httpServletRequest = (HttpServletRequest) request;
-			if (LOGGER.isInfoEnabled()) {
-				Enumeration<String> names = httpServletRequest.getHeaderNames();
-				while (names.hasMoreElements()) {
-					LOGGER.info("header:{}", names.nextElement());
-				}
-			}
-			final String username = httpServletRequest.getHeader(SEC_USERNAME);
-			if (username != null) {
-				SecurityContextHolder.getContext().setAuthentication(createAuthentication(httpServletRequest));
-
-				if (LOGGER.isInfoEnabled()) {
-					LOGGER.info("Populated SecurityContextHolder with pre-auth token: '{}'",
-							SecurityContextHolder.getContext().getAuthentication());
-				}
-			} else {
-				if (LOGGER.isInfoEnabled()) {
-					LOGGER.info("SecurityContextHolder not populated with pre-auth token");
-				}
-			}
-		}
-
-		chain.doFilter(request, response);
-	}
 
 	/**
 	 * Construction du token pre-authentification
@@ -148,4 +128,38 @@ public class PreAuthenticationFilter implements Filter {
 		return update;
 	}
 
+	@Override
+	protected void doFilterInternal(jakarta.servlet.http.HttpServletRequest request, HttpServletResponse response, jakarta.servlet.FilterChain filterChain) throws jakarta.servlet.ServletException, IOException {
+		if (request instanceof HttpServletRequest) {
+			HttpServletRequest httpServletRequest = request;
+			if (LOGGER.isInfoEnabled()) {
+				Enumeration<String> names = httpServletRequest.getHeaderNames();
+				while (names.hasMoreElements()) {
+					LOGGER.info("header:{}", names.nextElement());
+				}
+			}
+			final String username = httpServletRequest.getHeader(SEC_USERNAME);
+			if (username != null) {
+				SecurityContextHolder.getContext().setAuthentication(createAuthentication(httpServletRequest));
+
+				if (LOGGER.isInfoEnabled()) {
+					LOGGER.info("Populated SecurityContextHolder with pre-auth token: '{}'",
+							SecurityContextHolder.getContext().getAuthentication());
+				}
+			} else {
+				if (LOGGER.isInfoEnabled()) {
+					LOGGER.info("SecurityContextHolder not populated with pre-auth token");
+				}
+			}
+		}
+
+		filterChain.doFilter(request, response);
+	}
+	
+	@Override
+	protected boolean shouldNotFilter(final HttpServletRequest request) throws ServletException {
+		// Contrôle si l'URL n'est pas dans le liste d'exclusion. Si c'est le cas, elle
+		// ne passera pas dans ce filtre
+		return excludeUrlPatterns.stream().anyMatch(p -> pathMatcher.match(p, request.getServletPath()));
+	}
 }
